+++ tests/testit/test-format.R
@@ -96,3 +96,12 @@ assert('jsValuesHandleNull works', {
   (jsValuesHandleNull('abc') %==% jsValues('abc'))
 })
 
+assert('styleRow and styleEqual allows scalar values', {
+  result = styleRow(1:2, 'a')
+  expect = JS("$.inArray(dataIndex + 1, [1]) >= 0 ? \"a\" : $.inArray(dataIndex + 1, [2]) >= 0 ? \"a\" : null")
+  (result %==% expect)
+  result = styleEqual(1:2, 'a')
+  expect = JS("value == 1 ? \"a\" : value == 2 ? \"a\" : null")
+  (result %==% expect)
+})
+